refactor(test): migrate schedules and secrets helpers to three-tier architecture
- Create db-test-seeders/schedules.ts (3 seeders with @why-db-direct)
- Create db-test-assertions/schedules.ts (1 read-only assertion)
- Create db-test-seeders/secrets.ts (2 seeders with @why-db-direct)
- Remove DB functions from api-test-helpers/schedules.ts (keep 6 API
  functions + private helper)
- Remove DB functions from api-test-helpers/secrets.ts (keep 2 API
  functions)
- Add initServices() to all moved functions
- Update 5 consumer test files to import from correct tier
- No re-exports — direct imports consistent with established pattern

Closes #9369
