+++ turbo/apps/web/src/__tests__/api-test-helpers/secrets.ts
@@ -1,7 +1,3 @@
-import { secrets } from "../../db/schema/secret";
-import { variables } from "../../db/schema/variable";
-import { ORG_SENTINEL_USER_ID } from "../../lib/zero/org/org-sentinel";
-import { encryptSecretValue } from "../../lib/shared/crypto/secrets-encryption";
 import { POST as setSecretRoute } from "../../../app/api/zero/secrets/route";
 import { POST as setVariableRoute } from "../../../app/api/zero/variables/route";
 import { createTestRequest } from "./core";
@@ -85,34 +81,4 @@ export async function createTestVariable(
     );
   }
   return response.json();
-}
-
-export async function insertTestOrgSentinelSecret(params: {
-  orgId: string;
-  name: string;
-}): Promise<void> {
-  const { SECRETS_ENCRYPTION_KEY } = globalThis.services.env;
-  const encrypted = encryptSecretValue(
-    "sentinel-test-value",
-    SECRETS_ENCRYPTION_KEY,
-  );
-  await globalThis.services.db.insert(secrets).values({
-    name: params.name,
-    encryptedValue: encrypted,
-    type: "user",
-    userId: ORG_SENTINEL_USER_ID,
-    orgId: params.orgId,
-  });
-}
-
-export async function insertTestOrgSentinelVariable(params: {
-  orgId: string;
-  name: string;
-}): Promise<void> {
-  await globalThis.services.db.insert(variables).values({
-    name: params.name,
-    value: "sentinel-test-value",
-    userId: ORG_SENTINEL_USER_ID,
-    orgId: params.orgId,
-  });
 }
